LPS-81186 Move group accessibility validation to group local service to avoid cross reference between EIC SettingsMapFactoryImpl and SitesImpl
