+++ modules/apps/export-import/export-import-service/src/main/java/com/liferay/exportimport/internal/configuration/ExportImportConfigurationSettingsMapFactoryImpl.java
@@ -20,7 +20,6 @@ import com.liferay.exportimport.kernel.configuration.ExportImportConfigurationCo
 import com.liferay.exportimport.kernel.configuration.ExportImportConfigurationParameterMapFactoryUtil;
 import com.liferay.exportimport.kernel.configuration.ExportImportConfigurationSettingsMapFactory;
 import com.liferay.exportimport.kernel.lar.ExportImportHelper;
-import com.liferay.exportimport.kernel.staging.Staging;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Group;
@@ -29,6 +28,7 @@ import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.Http;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.UnicodeProperties;
@@ -268,7 +268,7 @@ public class ExportImportConfigurationSettingsMapFactoryImpl
 			portletRequest, "remoteAddress",
 			groupTypeSettingsProperties.getProperty("remoteAddress"));
 
-		remoteAddress = _staging.stripProtocolFromRemoteAddress(remoteAddress);
+		remoteAddress = _http.removeProtocol(remoteAddress);
 
 		int remotePort = ParamUtil.getInteger(
 			portletRequest, "remotePort",
@@ -288,7 +288,7 @@ public class ExportImportConfigurationSettingsMapFactoryImpl
 		boolean remotePrivateLayout = ParamUtil.getBoolean(
 			portletRequest, "remotePrivateLayout");
 
-		_staging.validateRemote(
+		_groupLocalService.validateRemote(
 			groupId, remoteAddress, remotePort, remotePathContext,
 			secureConnection, remoteGroupId);
 
@@ -400,6 +400,6 @@ public class ExportImportConfigurationSettingsMapFactoryImpl
 	private GroupLocalService _groupLocalService;
 
 	@Reference
-	private Staging _staging;
+	private Http _http;
 
 }
