+++ modules/apps/export-import/export-import-service/src/main/java/com/liferay/exportimport/internal/staging/StagingImpl.java
@@ -38,7 +38,6 @@ import com.liferay.exportimport.kernel.exception.LARFileSizeException;
 import com.liferay.exportimport.kernel.exception.LARTypeException;
 import com.liferay.exportimport.kernel.exception.LayoutImportException;
 import com.liferay.exportimport.kernel.exception.MissingReferenceException;
-import com.liferay.exportimport.kernel.exception.RemoteExportException;
 import com.liferay.exportimport.kernel.lar.ExportImportDateUtil;
 import com.liferay.exportimport.kernel.lar.ExportImportHelper;
 import com.liferay.exportimport.kernel.lar.ExportImportThreadLocal;
@@ -60,12 +59,10 @@ import com.liferay.portal.kernel.backgroundtask.BackgroundTask;
 import com.liferay.portal.kernel.backgroundtask.BackgroundTaskManager;
 import com.liferay.portal.kernel.exception.LayoutPrototypeException;
 import com.liferay.portal.kernel.exception.LocaleException;
-import com.liferay.portal.kernel.exception.NoSuchGroupException;
 import com.liferay.portal.kernel.exception.NoSuchLayoutBranchException;
 import com.liferay.portal.kernel.exception.NoSuchLayoutRevisionException;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.PortletIdException;
-import com.liferay.portal.kernel.exception.RemoteOptionsException;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
@@ -78,8 +75,6 @@ import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.model.BaseModel;
-import com.liferay.portal.kernel.model.ClassName;
-import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.LayoutBranch;
@@ -95,8 +90,6 @@ import com.liferay.portal.kernel.model.WorkflowInstanceLink;
 import com.liferay.portal.kernel.model.adapter.StagedTheme;
 import com.liferay.portal.kernel.scheduler.SchedulerEngineHelperUtil;
 import com.liferay.portal.kernel.security.auth.HttpPrincipal;
-import com.liferay.portal.kernel.security.auth.PrincipalException;
-import com.liferay.portal.kernel.security.auth.RemoteAuthException;
 import com.liferay.portal.kernel.security.pacl.DoPrivileged;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
@@ -128,7 +121,6 @@ import com.liferay.portal.kernel.util.Http;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringUtil;
@@ -141,7 +133,6 @@ import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.kernel.workflow.WorkflowTask;
 import com.liferay.portal.kernel.workflow.WorkflowTaskManagerUtil;
 import com.liferay.portal.kernel.xml.Element;
-import com.liferay.portal.service.http.ClassNameServiceHttp;
 import com.liferay.portal.service.http.GroupServiceHttp;
 import com.liferay.portal.util.PropsValues;
 import com.liferay.portlet.exportimport.service.http.StagingServiceHttp;
@@ -165,7 +156,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Objects;
 import java.util.ResourceBundle;
 import java.util.Set;
 
@@ -439,9 +429,9 @@ public class StagingImpl implements Staging {
 		boolean secureConnection = MapUtil.getBoolean(
 			settingsMap, "secureConnection");
 
-		validateRemoteGroup(
-			exportImportConfiguration.getGroupId(), targetGroupId,
-			remoteAddress, remotePort, remotePathContext, secureConnection);
+		_groupLocalService.validateRemote(
+			exportImportConfiguration.getGroupId(), remoteAddress, remotePort,
+			remotePathContext, secureConnection, targetGroupId);
 
 		boolean remotePrivateLayout = MapUtil.getBoolean(
 			settingsMap, "remotePrivateLayout");
@@ -507,9 +497,9 @@ public class StagingImpl implements Staging {
 			long remoteGroupId, boolean remotePrivateLayout)
 		throws PortalException {
 
-		validateRemoteGroup(
-			sourceGroupId, remoteGroupId, remoteAddress, remotePort,
-			remotePathContext, secureConnection);
+		_groupLocalService.validateRemote(
+			sourceGroupId, remoteAddress, remotePort, remotePathContext,
+			secureConnection, remoteGroupId);
 
 		PermissionChecker permissionChecker =
 			PermissionThreadLocal.getPermissionChecker();
@@ -2411,7 +2401,7 @@ public class StagingImpl implements Staging {
 				GetterUtil.getBoolean(
 					typeSettingsProperties.getProperty("secureConnection")));
 
-			validateRemote(
+			_groupLocalService.validateRemote(
 				sourceGroupId, remoteAddress, remotePort, remotePathContext,
 				secureConnection, sourceGroup.getRemoteLiveGroupId());
 
@@ -2707,7 +2697,7 @@ public class StagingImpl implements Staging {
 
 		remoteAddress = stripProtocolFromRemoteAddress(remoteAddress);
 
-		validateRemote(
+		_groupLocalService.validateRemote(
 			groupId, remoteAddress, remotePort, remotePathContext,
 			secureConnection, remoteGroupId);
 
@@ -2903,7 +2893,7 @@ public class StagingImpl implements Staging {
 			GetterUtil.getLong(
 				groupTypeSettingsProperties.getProperty("remoteGroupId")));
 
-		validateRemote(
+		_groupLocalService.validateRemote(
 			groupId, remoteAddress, remotePort, remotePathContext,
 			secureConnection, remoteGroupId);
 
@@ -3261,6 +3251,12 @@ public class StagingImpl implements Staging {
 		}
 	}
 
+	/**
+	 * @deprecated As of 5.0.0, replaced by {@link
+	 *             GroupLocalService#validateRemote(long, String, int, String,
+	 *             boolean, long)}
+	 */
+	@Deprecated
 	@Override
 	public void validateRemote(
 			long groupId, String remoteAddress, int remotePort,
@@ -3268,43 +3264,9 @@ public class StagingImpl implements Staging {
 			long remoteGroupId)
 		throws PortalException {
 
-		RemoteOptionsException roe = null;
-
-		if (!Validator.isDomain(remoteAddress) &&
-			!Validator.isIPAddress(remoteAddress)) {
-
-			roe = new RemoteOptionsException(
-				RemoteOptionsException.REMOTE_ADDRESS);
-
-			roe.setRemoteAddress(remoteAddress);
-
-			throw roe;
-		}
-
-		if ((remotePort < 1) || (remotePort > 65535)) {
-			roe = new RemoteOptionsException(
-				RemoteOptionsException.REMOTE_PORT);
-
-			roe.setRemotePort(remotePort);
-
-			throw roe;
-		}
-
-		if (Validator.isNotNull(remotePathContext) &&
-			(!remotePathContext.startsWith(StringPool.FORWARD_SLASH) ||
-			 remotePathContext.endsWith(StringPool.FORWARD_SLASH))) {
-
-			roe = new RemoteOptionsException(
-				RemoteOptionsException.REMOTE_PATH_CONTEXT);
-
-			roe.setRemotePathContext(remotePathContext);
-
-			throw roe;
-		}
-
-		validateRemoteGroup(
-			groupId, remoteGroupId, remoteAddress, remotePort,
-			remotePathContext, secureConnection);
+		_groupLocalService.validateRemote(
+			groupId, remoteAddress, remotePort, remotePathContext,
+			secureConnection, remoteGroupId);
 	}
 
 	/**
@@ -3580,17 +3542,6 @@ public class StagingImpl implements Staging {
 		return ParamUtil.getString(
 			portletRequest, param,
 			GetterUtil.getString(group.getTypeSettingsProperty(param)));
-	}
-
-	protected boolean isCompanyGroup(HttpPrincipal httpPrincipal, Group group) {
-		ClassName className = ClassNameServiceHttp.fetchByClassNameId(
-			httpPrincipal, group.getClassNameId());
-
-		if (Objects.equals(className.getClassName(), Company.class.getName())) {
-			return true;
-		}
-
-		return false;
 	}
 
 	protected boolean isLayoutRevisionIncomplete(
@@ -3953,128 +3904,6 @@ public class StagingImpl implements Staging {
 	@Deprecated
 	protected void setWorkflowInstanceLinkLocalService(
 		WorkflowInstanceLinkLocalService workflowInstanceLinkLocalService) {
-	}
-
-	protected void validateRemoteGroup(
-			long groupId, long remoteGroupId, String remoteAddress,
-			int remotePort, String remotePathContext, boolean secureConnection)
-		throws PortalException {
-
-		if (remoteGroupId <= 0) {
-			RemoteOptionsException roe = new RemoteOptionsException(
-				RemoteOptionsException.REMOTE_GROUP_ID);
-
-			roe.setRemoteGroupId(remoteGroupId);
-
-			throw roe;
-		}
-
-		Thread currentThread = Thread.currentThread();
-
-		ClassLoader contextClassLoader = currentThread.getContextClassLoader();
-
-		PermissionChecker permissionChecker =
-			PermissionThreadLocal.getPermissionChecker();
-
-		User user = permissionChecker.getUser();
-
-		String remoteURL = buildRemoteURL(
-			remoteAddress, remotePort, remotePathContext, secureConnection);
-
-		HttpPrincipal httpPrincipal = new HttpPrincipal(
-			remoteURL, user.getLogin(), user.getPassword(),
-			user.isPasswordEncrypted());
-
-		try {
-			currentThread.setContextClassLoader(
-				PortalClassLoaderUtil.getClassLoader());
-
-			// Ping the remote host and verify that the remote group exists in
-			// the same company as the remote user
-
-			GroupServiceHttp.checkRemoteStagingGroup(
-				httpPrincipal, remoteGroupId);
-
-			// Ensure that the local group and the remote group are not the same
-			// group and that they are either both company groups or both not
-			// company groups
-
-			Group group = _groupLocalService.getGroup(groupId);
-
-			Group remoteGroup = GroupServiceHttp.getGroup(
-				httpPrincipal, remoteGroupId);
-
-			if (group.isCompany() ^
-				isCompanyGroup(httpPrincipal, remoteGroup)) {
-
-				RemoteExportException ree = new RemoteExportException(
-					RemoteExportException.INVALID_GROUP);
-
-				ree.setGroupId(remoteGroupId);
-
-				throw ree;
-			}
-		}
-		catch (NoSuchGroupException nsge) {
-
-			// LPS-52675
-
-			if (_log.isDebugEnabled()) {
-				_log.debug(nsge, nsge);
-			}
-
-			RemoteExportException ree = new RemoteExportException(
-				RemoteExportException.NO_GROUP);
-
-			ree.setGroupId(remoteGroupId);
-
-			throw ree;
-		}
-		catch (PrincipalException pe) {
-
-			// LPS-52675
-
-			if (_log.isDebugEnabled()) {
-				_log.debug(pe, pe);
-			}
-
-			RemoteExportException ree = new RemoteExportException(
-				RemoteExportException.NO_PERMISSIONS);
-
-			ree.setGroupId(remoteGroupId);
-
-			throw ree;
-		}
-		catch (RemoteAuthException rae) {
-
-			// LPS-52675
-
-			if (_log.isDebugEnabled()) {
-				_log.debug(rae, rae);
-			}
-
-			rae.setURL(remoteURL);
-
-			throw rae;
-		}
-		catch (SystemException se) {
-
-			// LPS-52675
-
-			if (_log.isDebugEnabled()) {
-				_log.debug(se, se);
-			}
-
-			RemoteExportException ree = new RemoteExportException(
-				RemoteExportException.BAD_CONNECTION, se.getMessage());
-
-			ree.setURL(remoteURL);
-
-			throw ree;
-		}
-		finally {
-			currentThread.setContextClassLoader(contextClassLoader);
-		}
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(StagingImpl.class);
